move dataframes to package level; move data APIs (read and write) to blueprints
